requirements.txt
fitparse>=1.1.5
python-telegram-bot>=13.14,<14.0
